# storage,kv: remove Distinct from the Batch interface