--- a/pkg/kv/kvserver/abortspan/abortspan.go
+++ b/pkg/kv/kvserver/abortspan/abortspan.go
@@ -79,7 +79,7 @@ func (sc *AbortSpan) ClearData(e storage.Engine) error {
 	// NB: The abort span is a Range-ID local key which has no versions or intents.
 	iter := e.NewMVCCIterator(storage.MVCCKeyIterKind, storage.IterOptions{UpperBound: sc.max()})
 	defer iter.Close()
-	b := e.NewWriteOnlyBatch()
+	b := e.NewUnIndexedBatch(false /* supportReader */)
 	defer b.Close()
 	err := b.ClearIterRange(iter, sc.min(), sc.max())
 	if err != nil {