--- a/pkg/kv/kvserver/batcheval/cmd_conditional_put.go
+++ b/pkg/kv/kvserver/batcheval/cmd_conditional_put.go
@@ -51,15 +51,6 @@ func ConditionalPut(
 	if !args.Inline {
 		ts = h.Timestamp
 	}
-	if h.DistinctSpans {
-		if b, ok := readWriter.(storage.Batch); ok {
-			// Use the distinct batch for both blind and normal ops so that we don't
-			// accidentally flush mutations to make them visible to the distinct
-			// batch.
-			readWriter = b.Distinct()
-			defer readWriter.Close()
-		}
-	}
 
 	var expVal []byte
 	if len(args.ExpBytes) != 0 {